refactor(agent): localize attestation receiver adapter
